A bit more documentation. Made size return a long instead of an int in ServerBackedMap, just in case.

+++ src/main/java/org/mgenterprises/openbooks/customer/CustomerManager.java
@@ -69,7 +69,7 @@ public class CustomerManager extends ServerBackedMap<Customer>{
     }
 
     public int getHighestId() throws IOException {
-        return size();
+        return (int) size();
     }
     
 }
